refactor: eliminate unecessary polling render
