+++ app/components/Body.tsx
@@ -8,16 +8,18 @@ import HandsonTable from './HandsonTable'
 import { ApiAction } from '../store/api'
 
 import { Structure } from '../models/dataset'
-import { PageInfo } from '../models/store'
+import { PageInfo, WorkingDataset } from '../models/store'
 
 export interface BodyProps {
+  workingDataset: WorkingDataset
   peername: string
   name: string
   path: string
   value: any[]
   pageInfo: PageInfo
-  headers: any[]
-  onFetch: (page?: number, pageSize?: number) => Promise<ApiAction>
+  history: boolean
+  fetchBody: (page?: number, pageSize?: number) => Promise<ApiAction>
+  fetchCommitBody: (page?: number, pageSize?: number) => Promise<ApiAction>
   structure: Structure
 }
 
@@ -26,12 +28,38 @@ function shouldDisplayTable (value: any[], structure: Structure) {
   return value && structure && (structure.format === 'csv' || structure.format === 'xlsx' || structure.depth === 2)
 }
 
-const Body: React.FunctionComponent<BodyProps> = ({ value, pageInfo, headers, structure, onFetch }) => {
+const extractColumnHeaders = (workingDataset: WorkingDataset): undefined | object => {
+  const schema = workingDataset.components.schema.value
+
+  if (!schema) {
+    return undefined
+  }
+
+  if (schema && (!schema.items || (schema.items && !schema.items.items))) {
+    return undefined
+  }
+
+  return schema && schema.items && schema.items.items.map((d: { title: string }): string => d.title)
+}
+
+const Body: React.FunctionComponent<BodyProps> = (props) => {
+  const {
+    value,
+    pageInfo,
+    structure,
+    workingDataset,
+    history,
+    fetchBody,
+    fetchCommitBody
+  } = props
   const isLoadingFirstPage = (pageInfo.page === 1 && pageInfo.isFetching)
 
   const handleScrollToBottom = () => {
+    const onFetch = history ? fetchCommitBody : fetchBody
     onFetch(pageInfo.page + 1, pageInfo.pageSize)
   }
+
+  const headers = extractColumnHeaders(workingDataset)
 
   return (
     <div className='transition-group'>
