+++ package.json
@@ -6,7 +6,7 @@
   "main": "main.js",
   "scripts": {
     "codecov": "./node_modules/.bin/codecov",
-    "test": "jest --testPathIgnorePatterns=.*\/e2e\/.* --coverage",
+    "test": "jest --testPathIgnorePatterns=.*/e2e/.* --coverage",
     "test-all": "npm run test && npm run build && npm run test-e2e",
     "test-watch": "npm test -- --watch",
     "test-e2e": "cross-env NODE_ENV=test node --trace-warnings ./test/runTests.js e2e",
@@ -118,7 +118,6 @@
   ],
   "homepage": "https://qri.io",
   "devDependencies": {
-    "codecov": "3.5.0",
     "@types/classnames": "2.2.9",
     "@types/enzyme": "3.10.2",
     "@types/history": "4.7.2",
@@ -141,6 +140,7 @@
     "@typescript-eslint/eslint-plugin": "1.11.0",
     "@typescript-eslint/parser": "1.11.0",
     "asar": "2.0.1",
+    "codecov": "3.5.0",
     "concurrently": "4.1.2",
     "cross-env": "5.2.0",
     "css-loader": "3.0.0",
@@ -203,6 +203,7 @@
     "@types/webpack-env": "^1.14.0",
     "babel-loader": "^8.0.6",
     "classnames": "2.2.6",
+    "deep-equal": "1.1.0",
     "electron-debug": "3.0.1",
     "electron-updater": "4.1.2",
     "eslint-utils": "1.4.2",
